Split scraper into two new modules
Moved the website get methods to getInflation.py
